% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{get_account}
\alias{get_account}
\title{AWS Lambda Account Settings}
\usage{
get_account(...)
}
\arguments{
\item{\dots}{Additional arguments passed to \code{\link{lambdaHTTP}}.}
}
\value{
A list.
}
\description{
Get account settings
}
\examples{
\dontrun{
  get_account()
}
}
