% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_function_name}
\alias{get_function_name}
\alias{get_function_name.character}
\alias{get_function_name.aws_lambda_function}
\title{Get name of Lambda function}
\usage{
get_function_name(x, ...)

\method{get_function_name}{character}(x, ...)

\method{get_function_name}{aws_lambda_function}(x, ...)
}
\arguments{
\item{x}{An object of class \dQuote{aws_lambda_function}.}

\item{\dots}{Additional arguments passed to methods}
}
\description{
Extracts an AWS Lambda function's name
}
