% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{op-null-default}
\alias{op-null-default}
\alias{\%||\%}
\title{Default value for `NULL`}
\usage{
x \%||\% y
}
\arguments{
\item{x, y}{If `x` is NULL, will return `y`; otherwise returns `x`.}
}
\description{
This infix function makes it easy to replace `NULL`s with a default
value. It's inspired by the way that Ruby's or operation (`||`)
works. Copied from the rlang package.
}
\keyword{internal}
